Add the base CMake build
This is the CMakeLists.txt and supporting file that allow using CMake to
build FSON and install to the user's CMake registry.  After building and
installing, (with a possible CMAKE_INSTALL_PREFIX=...), the user can
simply add

    find_package(FSON)
    target_link_libraries(mytarget FSON::FSON)

and everything will Just Work™.

Note, this only deals with building the library and does not yet build
the tests.

[FetchContent]: https://cmake.org/cmake/help/latest/module/FetchContent.html