Add the unit tests to the CMake build
After taking some more time to review the Meson build, I deduced how to
include the tests with the CMake build.  We have to add [zofu][zofu],
and we need to tell CMake how to build it.  The building of zofu could
be replaced by the appropriate CMake import details if it gets a CMake
build.

[zofu]: https://github.com/acroucher/zofu